pyproject.toml
[tool.poetry]
name = "rumahku"
version = "0.1.0"
description = "Service to support semantic search over e-commerce ctalog"
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
litestar = {extras = ["standard"], version = "^2.2.1"}
# litestar = { extras = ["standard"], path = "../litestar/", develop = true }
weaviate-client = "^3.25.2"
python-dotenv = "^1.0.0"
pandas = "^2.1.2"

[tool.poetry.dev-dependencies]
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
litestar = { version = "*", extras = ["testing"] }
# litestar = { extras = ["testing"], path  = "../litestar/", develop = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.coverage.report]
show_missing = true
exclude_lines = [
    'if TYPE_CHECKING:',
    'pragma: no cover',
    "if __name__ == '__main__':",
]

[tool.coverage.run]
omit = [
    "tests/*",
]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = false
implicit_reexport = false
show_error_codes = true
strict = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
addopts = "--cov=app -v"
